simple and fong

+++ src/Point2D.java
@@ -3,19 +3,19 @@ import java.awt.Color;
 
 public class Point2D {
 
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	Color color;
 	
 	public Point2D(){}
 	
-	public Point2D(double x, double y){
+	public Point2D(int x, int y){
 		this.x = x;
 		this.y = y;
 	}
 	
 	
-	public Point2D(double x, double y, Color color){
+	public Point2D(int x, int y, Color color){
 		this.x = x;
 		this.y = y;
 		this.color = color;
@@ -25,16 +25,16 @@ public class Point2D {
 		return color;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
-	public double getY() {
+	public int getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
 }
